many updates, the multipop model is much stabler now and passes tests where there are many populations but all individuals come from the same pop- this returns same arg as a single (or no-population) model with the same random number seed.
The biggest change in this version was changing the way lineages were discounted
in count_lienages_internal for times < minage. This does not work in a multipop
setting and instead counts have to be adjusted on a case-by-case basis throughout
the code.

Another small but significant change was renaming the SubPath class to UniquePath.